# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CRM Proto API",
    "version": "1.0.0",
    "description": "# CRM Proto API\n\nREST API для CRM системы.\n\n## Аутентификация\n\nAPI использует cookie-based аутентификацию через NextAuth.js.\nДля доступа к защищённым эндпоинтам необходимо авторизоваться через `/login`.\n\n## Ошибки\n\nВсе ошибки возвращаются в формате:\n\n```json\n{\n  \"error\": \"Описание ошибки\"\n}\n```\n\nСтандартные коды ответов:\n- `200` - Успешный запрос\n- `201` - Ресурс создан\n- `400` - Неверный запрос (ошибка валидации)\n- `401` - Не авторизован\n- `403` - Доступ запрещён\n- `404` - Ресурс не найден\n- `500` - Внутренняя ошибка сервера"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/api/opportunities/search": {
      "post": {
        "summary": "Поиск сделок",
        "description": "Возвращает список сделок с пагинацией и фильтрацией. Фильтры передаются в теле запроса.",
        "tags": [
          "Сделки"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "opportunities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Уникальный идентификатор"
                          },
                          "title": {
                            "type": "string",
                            "description": "Название"
                          },
                          "description": {
                            "type": "string"
                          },
                          "value": {
                            "type": "number",
                            "description": "Сумма"
                          },
                          "currency": {
                            "type": "string",
                            "description": "Валюта"
                          },
                          "stage": {
                            "type": "string",
                            "enum": [
                              "prospecting",
                              "qualification",
                              "proposal",
                              "negotiation",
                              "closed_won",
                              "closed_lost"
                            ],
                            "description": "Стадия"
                          },
                          "priority": {
                            "type": "string",
                            "enum": [
                              "low",
                              "medium",
                              "high"
                            ],
                            "description": "Приоритет"
                          },
                          "probability": {
                            "type": "number",
                            "description": "Вероятность закрытия"
                          },
                          "expectedCloseDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "actualCloseDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "contactId": {
                            "type": "string"
                          },
                          "ownerId": {
                            "type": "string",
                            "description": "ID владельца"
                          },
                          "notes": {
                            "type": "string"
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "description": "Теги"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Дата создания"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Дата обновления"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "value",
                          "currency",
                          "stage",
                          "priority",
                          "probability",
                          "ownerId",
                          "tags",
                          "createdAt",
                          "updatedAt"
                        ],
                        "description": "Сделка"
                      },
                      "description": "Список сделок"
                    },
                    "total": {
                      "type": "number",
                      "description": "Общее количество"
                    },
                    "page": {
                      "type": "number",
                      "description": "Текущая страница"
                    },
                    "limit": {
                      "type": "number",
                      "description": "Размер страницы"
                    }
                  },
                  "required": [
                    "opportunities",
                    "total",
                    "page",
                    "limit"
                  ],
                  "description": "Список сделок с пагинацией"
                },
                "example": {
                  "data": [
                    {
                      "id": "clx1234567890abcdef",
                      "title": "Внедрение CRM системы",
                      "description": "Проект по внедрению CRM для автоматизации продаж",
                      "amount": 1500000,
                      "currency": "USD",
                      "stage": "negotiation",
                      "priority": "high",
                      "probability": 70,
                      "expectedCloseDate": "2024-03-15",
                      "contactId": "clx0987654321fedcba",
                      "ownerId": "clxowner123456789ab",
                      "contact": {
                        "id": "clx0987654321fedcba",
                        "firstName": "Иван",
                        "lastName": "Петров",
                        "company": "ООО \"Технологии\""
                      },
                      "createdAt": "2024-01-20T14:00:00.000Z",
                      "updatedAt": "2024-02-01T09:30:00.000Z"
                    }
                  ],
                  "total": 1,
                  "page": 1,
                  "limit": 20,
                  "totalPages": 1
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "search": {
                    "type": "string"
                  },
                  "stage": {
                    "type": "string",
                    "enum": [
                      "prospecting",
                      "qualification",
                      "proposal",
                      "negotiation",
                      "closed_won",
                      "closed_lost"
                    ],
                    "description": "Стадия сделки: prospecting - поиск, qualification - квалификация, proposal - предложение, negotiation - переговоры, closed_won - выиграна, closed_lost - проиграна"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high"
                    ],
                    "description": "Приоритет: low - низкий, medium - средний, high - высокий"
                  },
                  "ownerId": {
                    "type": "string"
                  },
                  "contactId": {
                    "type": "string"
                  },
                  "minValue": {
                    "type": "number"
                  },
                  "maxValue": {
                    "type": "number"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "page": {
                    "type": "integer",
                    "minimum": 0,
                    "default": 1
                  },
                  "limit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100,
                    "default": 20
                  }
                },
                "description": "Фильтры для списка сделок"
              }
            }
          }
        }
      }
    },
    "/api/opportunities": {
      "post": {
        "summary": "Создать сделку",
        "description": "Создаёт новую сделку. Владельцем автоматически становится текущий пользователь.",
        "tags": [
          "Сделки"
        ],
        "responses": {
          "200": {
            "description": "Успешный ответ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Уникальный идентификатор"
                    },
                    "title": {
                      "type": "string",
                      "description": "Название"
                    },
                    "description": {
                      "type": "string"
                    },
                    "value": {
                      "type": "number",
                      "description": "Сумма"
                    },
                    "currency": {
                      "type": "string",
                      "description": "Валюта"
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "prospecting",
                        "qualification",
                        "proposal",
                        "negotiation",
                        "closed_won",
                        "closed_lost"
                      ],
                      "description": "Стадия"
                    },
                    "priority": {
                      "type": "string",
                      "enum": [
                        "low",
                        "medium",
                        "high"
                      ],
                      "description": "Приоритет"
                    },
                    "probability": {
                      "type": "number",
                      "description": "Вероятность закрытия"
                    },
                    "expectedCloseDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "actualCloseDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "contactId": {
                      "type": "string"
                    },
                    "ownerId": {
                      "type": "string",
                      "description": "ID владельца"
                    },
                    "notes": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Теги"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата создания"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Дата обновления"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "value",
                    "currency",
                    "stage",
                    "priority",
                    "probability",
                    "ownerId",
                    "tags",
                    "createdAt",
                    "updatedAt"
                  ],
                  "description": "Сделка"
                },
                "example": {
                  "id": "clx1234567890abcdef",
                  "title": "Внедрение CRM системы",
                  "description": "Проект по внедрению CRM для автоматизации продаж",
                  "amount": 1500000,
                  "currency": "USD",
                  "stage": "negotiation",
                  "priority": "high",
                  "probability": 70,
                  "expectedCloseDate": "2024-03-15",
                  "contactId": "clx0987654321fedcba",
                  "ownerId": "clxowner123456789ab",
                  "contact": {
                    "id": "clx0987654321fedcba",
                    "firstName": "Иван",
                    "lastName": "Петров",
                    "company": "ООО \"Технологии\""
                  },
                  "createdAt": "2024-01-20T14:00:00.000Z",
                  "updatedAt": "2024-02-01T09:30:00.000Z"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Название сделки",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "value": {
                    "type": "number",
                    "description": "Сумма сделки",
                    "minimum": 0
                  },
                  "currency": {
                    "type": "string"
                  },
                  "stage": {
                    "type": "string",
                    "enum": [
                      "prospecting",
                      "qualification",
                      "proposal",
                      "negotiation",
                      "closed_won",
                      "closed_lost"
                    ],
                    "description": "Стадия сделки: prospecting - поиск, qualification - квалификация, proposal - предложение, negotiation - переговоры, closed_won - выиграна, closed_lost - проиграна"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "high"
                    ],
                    "description": "Приоритет: low - низкий, medium - средний, high - высокий"
                  },
                  "probability": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "expectedCloseDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "contactId": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "title",
                  "value"
                ],
                "description": "Данные для создания сделки"
              },
              "example": {
                "title": "Внедрение CRM системы",
                "description": "Проект по внедрению CRM для автоматизации продаж",
                "amount": 1500000,
                "currency": "USD",
                "stage": "qualification",
                "priority": "high",
                "probability": 30,
                "expectedCloseDate": "2024-03-15",
                "contactId": "clx0987654321fedcba"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}